+++ docs/CHANGELOG.md
@@ -15,6 +15,7 @@ Nos apegamos a [SEMVER](SEMVER.md), revisa la información para entender mejor e
 ## Version UNRELEASED
 
 - Se agrega el método `StampingAlert::extraInfo()` para obtener la respuesta de la incidencia en `ExtraInfo`.
+- Se agrega el método `StampingResult::faultCode()` para obtener la respuesta en `faultcode`.
 
 ## Version 0.2.5 2020-01-14